Remove Extra Controllers and update connection string
