+++ DemoApp/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Server=CHARUL;Database=DemoApp;Trusted_Connection=True;MultipleActiveResultSets=true"
+    "DefaultConnection": "Server=CHARUL;Database=DemoApp;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=true"
   },
 
   "Logging": {
